# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NationsEpoque API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface NationSummary {
  id: string;
  name: string;
  description?: string | null;
  leaderId: string;
  level: number;
  xp: number;
  seasonPoints: number;
  bankBalance: number;
  open: boolean;
  coalitionId?: string | null;
  createdAt: number;
  memberCount: number;
}

export interface NationMember {
  playerId: string;
  nationId: string;
  roleName: string;
  playerName?: string | null;
  joinedAt: number;
}

export type NationDetail = NationSummary & {
  members: NationMember[];
  allies: string[];
};

export interface ClaimedChunk {
  id: string;
  nationId: string;
  worldName: string;
  chunkX: number;
  chunkZ: number;
  claimedAt: number;
}

export interface War {
  id: string;
  attackerNationId: string;
  defenderNationId: string;
  type: string;
  status: string;
  declaredAt: number;
  endsAt: number;
  reason?: string | null;
  attackerKills: number;
  defenderKills: number;
  staffNote?: string | null;
  validatedBy?: string | null;
  surrenderRequested: boolean;
  surrenderRequestedAt?: number | null;
}

export interface CoalitionDetail {
  id: string;
  name: string;
  leaderNationId: string;
  createdAt: number;
  memberNationIds: string[];
}

export interface PlayerProfile {
  playerId: string;
  playerName?: string | null;
  balance?: number | null;
  grade?: string | null;
  gradeLevel?: number | null;
  gradeXp?: number | null;
  totalPlaytimeSeconds?: number | null;
  claimedPlaytimeSeconds?: number | null;
}

export interface SeasonStat {
  playerId: string;
  seasonNumber: number;
  playerName?: string | null;
  kills: number;
  deaths: number;
  warsWon: number;
  claims: number;
}

export interface Season {
  seasonNumber: number;
  startedAt: number;
  endedAt?: number | null;
  current: boolean;
}

export type ListNationsParams = {
  limit?: number;
  offset?: number;
  sortBy?: ListNationsSortBy;
  order?: ListNationsOrder;
};

export type ListNationsSortBy =
  (typeof ListNationsSortBy)[keyof typeof ListNationsSortBy];

export const ListNationsSortBy = {
  name: "name",
  level: "level",
  seasonPoints: "seasonPoints",
  bankBalance: "bankBalance",
  memberCount: "memberCount",
} as const;

export type ListNationsOrder =
  (typeof ListNationsOrder)[keyof typeof ListNationsOrder];

export const ListNationsOrder = {
  asc: "asc",
  desc: "desc",
} as const;

export type ListNations200 = {
  data: NationSummary[];
  total: number;
};

export type GetNationWarsParams = {
  status?: GetNationWarsStatus;
};

export type GetNationWarsStatus =
  (typeof GetNationWarsStatus)[keyof typeof GetNationWarsStatus];

export const GetNationWarsStatus = {
  PENDING: "PENDING",
  ACTIVE: "ACTIVE",
  FINISHED: "FINISHED",
  CANCELLED: "CANCELLED",
} as const;

export type ListWarsParams = {
  status?: ListWarsStatus;
  limit?: number;
  offset?: number;
};

export type ListWarsStatus =
  (typeof ListWarsStatus)[keyof typeof ListWarsStatus];

export const ListWarsStatus = {
  PENDING: "PENDING",
  ACTIVE: "ACTIVE",
  FINISHED: "FINISHED",
  CANCELLED: "CANCELLED",
} as const;

export type ListWars200 = {
  data: War[];
  total: number;
};

export type GetSeasonStatsParams = {
  limit?: number;
  offset?: number;
};

export type GetSeasonStats200 = {
  data: SeasonStat[];
  total: number;
};

export type GetNationsLeaderboardParams = {
  limit?: number;
};

export type GetPlayersLeaderboardParams = {
  limit?: number;
};