# lib/api-zod/src/generated/types/coalitionDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NationsEpoque API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CoalitionDetail {
  id: string;
  name: string;
  leaderNationId: string;
  createdAt: number;
  memberNationIds: string[];
}