# lib/api-zod/src/generated/types/war.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NationsEpoque API specification
 * OpenAPI spec version: 0.1.0
 */

export interface War {
  id: string;
  attackerNationId: string;
  defenderNationId: string;
  type: string;
  status: string;
  declaredAt: number;
  endsAt: number;
  reason?: string | null;
  attackerKills: number;
  defenderKills: number;
  staffNote?: string | null;
  validatedBy?: string | null;
  surrenderRequested: boolean;
  surrenderRequestedAt?: number | null;
}